Player automatically owner

+++ src/fr/aphaistos/homes/json/HomePos.java
@@ -5,38 +5,38 @@ import java.util.Map.Entry;
 import com.google.gson.internal.LinkedTreeMap;
 
 public class HomePos {
-	private final int x;
-	private final int y;
-	private final int z;
+	private final double x;
+	private final double y;
+	private final double z;
 	
-	public HomePos(final int x, final int y, final int z) {
+	public HomePos(final double x, final double y, final double z) {
 		super();
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
-	public int getY() {
+	public double getY() {
 		return y;
 	}
-	public int getZ() {
+	public double getZ() {
 		return z;
 	}
 
 	public static HomePos fromLinkedTreeMap(LinkedTreeMap<String, Object> map) {
-		int x = 0;
-		int y = 0;
-		int z = 0;
+		double x = 0;
+		double y = 0;
+		double z = 0;
 		for (Entry<String, Object> entry : map.entrySet())
 			if (entry.getKey().equals("x"))
-				x = (int) ((double) entry.getValue());
+				x = (double) entry.getValue();
 			else if (entry.getKey().equals("y"))
-				y = (int) ((double) entry.getValue());
+				y = (double) entry.getValue();
 			else if (entry.getKey().equals("z"))
-				z = (int) ((double) entry.getValue());
+				z = (double) entry.getValue();
 		return new HomePos(x, y, z);
 	}
 }
